# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProxyAdminPanel",
    "description": "X-UI style proxy management system with BFF architecture",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/core-config/current": {
      "get": {
        "tags": [
          "Core Configuration"
        ],
        "summary": "Get Core Config",
        "description": "Get current Core Service configuration.",
        "operationId": "get_core_config_api_core_config_current_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreConfigResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/core-config/update": {
      "post": {
        "tags": [
          "Core Configuration"
        ],
        "summary": "Update Core Config",
        "description": "Update Core Service configuration.\nUpdates the .env file with new settings.",
        "operationId": "update_core_config_api_core_config_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoreConfigModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/core-config/test": {
      "post": {
        "tags": [
          "Core Configuration"
        ],
        "summary": "Test Core Connection",
        "description": "Test connection to Core Service.\nIf config is provided, test with those settings.\nOtherwise, test with current settings from .env.",
        "operationId": "test_core_connection_api_core_config_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CoreConfigModel"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Config"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/api-keys/": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List Api Keys",
        "description": "List all API keys.",
        "operationId": "list_api_keys_api_api_keys__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/APIKeyResponse"
                  },
                  "type": "array",
                  "title": "Response List Api Keys Api Api Keys  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key for external system integration.\nReturns the full API key - save it securely as it won't be shown again.",
        "operationId": "create_api_key_api_api_keys__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyWithSecret"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/api-keys/{key_id}": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "Get Api Key",
        "description": "Get API key by ID.",
        "operationId": "get_api_key_api_api_keys__key_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "API Keys"
        ],
        "summary": "Delete Api Key",
        "description": "Delete an API key.",
        "operationId": "delete_api_key_api_api_keys__key_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/api-keys/{key_id}/toggle": {
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Toggle Api Key",
        "description": "Toggle API key active status.",
        "operationId": "toggle_api_key_api_api_keys__key_id__toggle_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "can_read": {
            "type": "boolean",
            "title": "Can Read",
            "default": true
          },
          "can_write": {
            "type": "boolean",
            "title": "Can Write",
            "default": false
          },
          "can_delete": {
            "type": "boolean",
            "title": "Can Delete",
            "default": false
          },
          "rate_limit_per_minute": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Rate Limit Per Minute",
            "default": 60
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "APIKeyCreate"
      },
      "APIKeyResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "can_read": {
            "type": "boolean",
            "title": "Can Read"
          },
          "can_write": {
            "type": "boolean",
            "title": "Can Write"
          },
          "can_delete": {
            "type": "boolean",
            "title": "Can Delete"
          },
          "rate_limit_per_minute": {
            "type": "integer",
            "title": "Rate Limit Per Minute"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "key_prefix",
          "is_active",
          "can_read",
          "can_write",
          "can_delete",
          "rate_limit_per_minute",
          "last_used_at",
          "expires_at",
          "created_at",
          "updated_at"
        ],
        "title": "APIKeyResponse"
      },
      "APIKeyWithSecret": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "can_read": {
            "type": "boolean",
            "title": "Can Read"
          },
          "can_write": {
            "type": "boolean",
            "title": "Can Write"
          },
          "can_delete": {
            "type": "boolean",
            "title": "Can Delete"
          },
          "rate_limit_per_minute": {
            "type": "integer",
            "title": "Rate Limit Per Minute"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "key_prefix",
          "is_active",
          "can_read",
          "can_write",
          "can_delete",
          "rate_limit_per_minute",
          "last_used_at",
          "expires_at",
          "created_at",
          "updated_at",
          "api_key"
        ],
        "title": "APIKeyWithSecret",
        "description": "Response when creating a new API key - includes the full key"
      },
...
      "ConnectionTestResult": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "api_url": {
            "type": "string",
            "title": "Api Url"
          },
          "response_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Time Ms"
          },
          "interfaces": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interfaces"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "api_url"
        ],
        "title": "ConnectionTestResult",
        "description": "Connection test result."
      },
      "CoreConfigModel": {
        "properties": {
          "api_url": {
            "type": "string",
            "title": "Api Url"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "api_url",
          "api_key"
        ],
        "title": "CoreConfigModel",
        "description": "Core Service configuration model."
      },
      "CoreConfigResponse": {
        "properties": {
          "api_url": {
            "type": "string",
            "title": "Api Url"
          },
          "api_key_masked": {
            "type": "string",
            "title": "Api Key Masked"
          },
          "is_configured": {
            "type": "boolean",
            "title": "Is Configured"
          }
        },
        "type": "object",
        "required": [
          "api_url",
          "api_key_masked",
          "is_configured"
        ],
        "title": "CoreConfigResponse",
        "description": "Core Service configuration response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}